requirements.txt
PyYAML>=3.10
SQLAlchemy>=1.0
discord.py>=1.5
psycopg2>=2.7
alembic>=1.8
